Refactor pagination a bit to optimize requests; Optimize SEO;
