+++ frontend/components/search/PaginationCity.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { countries as supportedCountries } from '../../../common/const/supportedCountries';
 import { useRoute } from 'vue-router';
+import { CommonErrorsEnum } from '../../../common/const';
 
 const { t, locale } = useI18n();
 
@@ -76,6 +77,26 @@ const {
 		},
 		watch: [tags.value, dateStart.value, dateEnd.value, currentPage.value]
 	}
+});
+
+if (errorEvents.value) {
+	if (
+		errorEvents.value.data &&
+		errorEvents.value.data.message === CommonErrorsEnum.CITY_NOT_FOUND
+	) {
+		throw createError({
+			statusCode: 404,
+			data: { message: t(`errors.${CommonErrorsEnum.CITY_NOT_FOUND}`, { city: city }) }
+		});
+	}
+}
+
+useHead({
+	script: [
+		posterEvents.value
+			? getJSONEventList(posterEvents.value.docs, locale.value, route.path)
+			: undefined
+	]
 });
 
 watch(
@@ -201,6 +222,14 @@ watch(
 					/>
 				</NuxtLink>
 			</div>
+			<SearchLoader
+				v-if="pending"
+				:size="mobile ? 'middle' : 'big'"
+			/>
+			<SearchNotFound
+				v-if="!pending && !posterEvents"
+				:title="$t('event.filteredEvents.no_events_found')"
+			/>
 			<SearchEventCardsList
 				v-if="posterEvents.docs && posterEvents.docs.length !== 0"
 				:events="posterEvents.docs"
