+++ frontend/i18n/locales/ru-RU/meta.ts
@@ -2,9 +2,10 @@ export default {
 	site_name: 'Афиша Peredelano',
 	default_title: {
 		first: 'Афиша мероприятий Сербии',
-		second: 'Афиша мероприятий Черногории - Peredelano Afisha'
+		second: 'Афиша мероприятий Черногории - Афиша Переделано'
 	},
-
+	default_keywords:
+		'лучшие мероприятия Сербии, лучшие мероприятия Черногории, Белград, Цетине, огромный выбор мероприятий, афиша событий 2024, расписание, билеты, мероприятия, выставки, экскурсии, шоу, концерты, рок-концерты, фестивали, спектакли, флешмобы, вечеринки, детям, схема, адрес, фото',
 	default_description:
 		'В каталоге представлены десятки мероприятий Сербии, Черногории по различным направлениям: концерты, встречи, выставки, фестивали, мастер-классы',
 	about_us: {
